Soporte H2

+++ src/main/java/ec/cjpq/academicoboot/repository/ActividadRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Actividad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActividadRepository extends JpaRepository<Actividad, Integer>{
+public interface ActividadRepository extends JpaRepository<Actividad, Long>{
 }
