+++ src/main/java/ec/cjpq/academicoboot/repository/AlumnoRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer>{
+public interface AlumnoRepository extends JpaRepository<Alumno, Long>{
 }
