+++ src/main/java/ec/cjpq/academicoboot/repository/CalendarioEspolRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.CalendarioEspol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarioEspolRepository extends JpaRepository<CalendarioEspol, Integer>{
+public interface CalendarioEspolRepository extends JpaRepository<CalendarioEspol, Long>{
 }
