+++ src/main/java/ec/cjpq/academicoboot/repository/MateriaRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Materia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MateriaRepository extends JpaRepository<Materia, Integer>{
+public interface MateriaRepository extends JpaRepository<Materia, Long>{
 }
