+++ src/main/java/ec/cjpq/academicoboot/repository/NivelAcademicoRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.NivelAcademico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NivelAcademicoRepository extends JpaRepository<NivelAcademico, Integer>{
+public interface NivelAcademicoRepository extends JpaRepository<NivelAcademico, Long>{
 }
