+++ src/main/java/ec/cjpq/academicoboot/repository/ParaleloRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Paralelo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParaleloRepository extends JpaRepository<Paralelo, Integer>{
+public interface ParaleloRepository extends JpaRepository<Paralelo, Long>{
 }
