+++ src/main/java/ec/cjpq/academicoboot/repository/ProfesorRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Profesor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfesorRepository extends JpaRepository<Profesor, Integer>{
+public interface ProfesorRepository extends JpaRepository<Profesor, Long>{
 }
