+++ src/main/java/ec/cjpq/academicoboot/repository/RegistroRepository.java
@@ -3,6 +3,6 @@ package ec.cjpq.academicoboot.repository;
 import ec.cjpq.academicoboot.model.Registro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegistroRepository extends JpaRepository<Registro, Integer>{
+public interface RegistroRepository extends JpaRepository<Registro, Long>{
 }
